Add: Not clear inl

+++ clear.sh
@@ -2,6 +2,13 @@
 cd /home/origami/code/bisheng/SysY2021/src
 
 rm $(find . | grep ".*\.o")
-rm $(find . | grep ".*\.inl")
-rm frontend/parser.tab.c frontend/parser.tab.h frontend/lex.yy.c
+
+if [ $1 == "no-gen" ]
+then
+    echo "Warning: No generated file is cleaned."
+else
+    rm $(find . | grep ".*\.inl")
+    rm frontend/parser.tab.c frontend/parser.tab.h frontend/lex.yy.c
+fi
+
 rm ../ssyc
